Move code to modules, add prelude, remove libc
