+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name          = "opengles"
-version       = "0.2.1"
+version       = "0.3.0"
 authors       = ["Sean Kerr <sean@metatomic.io>", "Lauri Gustafsson <me@gustafla.space>"]
 license       = "Apache-2.0"
 description   = "Rust bindings for OpenGL ES 2.0"
@@ -12,4 +12,3 @@ exclude       = [".gitignore"]
 
 [dependencies]
 khronos = "0.1.2"
-libc    = "0.1.10"
