+++ src/glesv2/types.rs
@@ -1,0 +1,20 @@
+use super::*;
+
+pub type GLbitfield = c_uint;
+pub type GLboolean = c_uchar;
+pub type GLbyte = khronos_int8_t;
+pub type GLchar = c_char;
+pub type GLclampf = khronos_float_t;
+pub type GLenum = c_uint;
+pub type GLfixed = khronos_int32_t;
+pub type GLfloat = khronos_float_t;
+pub type GLint = c_int;
+pub type GLshort = c_short;
+pub type GLsizei = c_int;
+pub type GLubyte = khronos_uint8_t;
+pub type GLuint = c_uint;
+pub type GLushort = c_ushort;
+pub type GLvoid = c_void;
+
+pub type GLintptr = khronos_intptr_t;
+pub type GLsizeiptr = khronos_ssize_t;
