+++ src/lib.rs
@@ -23,10 +23,19 @@
 // -------------------------------------------------------------------------------------------------
 
 extern crate khronos;
-extern crate libc;
 
 // -------------------------------------------------------------------------------------------------
 // MODULES
 // -------------------------------------------------------------------------------------------------
 
 pub mod glesv2;
+
+// -------------------------------------------------------------------------------------------------
+// PRELUDE
+// -------------------------------------------------------------------------------------------------
+
+pub mod prelude {
+    pub use glesv2;
+    pub use glesv2::constants::*;
+    pub use glesv2::types::*;
+}
